+++ api/src/main/java/org/allaymc/api/entity/interfaces/EntityEnderCrystal.java
@@ -1,7 +1,7 @@
 package org.allaymc.api.entity.interfaces;
 
-import org.allaymc.api.entity.Entity;
+import org.allaymc.api.entity.component.EntityEnderCrystalBaseComponent;
 
-public interface EntityEnderCrystal extends Entity {
+public interface EntityEnderCrystal extends EntityLiving, EntityEnderCrystalBaseComponent {
 
 }