+++ server/src/main/java/org/allaymc/server/player/AllayPlayer.java
@@ -557,6 +557,9 @@ public class AllayPlayer implements Player {
         return map;
     }
 
+    /**
+     * Adds generic metadata for the specified {@code entity} to the provided {@code map}.
+     */
     protected void addGenericMetadata(Entity entity, EntityDataMap map) {
         map.setFlag(EntityFlag.HAS_COLLISION, entity.hasEntityCollision());
         map.setFlag(EntityFlag.CAN_CLIMB, true);
@@ -591,6 +594,10 @@ public class AllayPlayer implements Player {
         }
     }
 
+    /**
+     * Adds component-specific metadata to the {@code map} for the given {@code entity}. This method
+     * applies metadata based on its specific components.
+     */
     protected void addComponentSpecificMetadata(Entity entity, EntityDataMap map) {
         if (entity instanceof EntityPhysicsComponent physicsComponent) {
             map.setFlag(EntityFlag.HAS_GRAVITY, physicsComponent.hasGravity());
@@ -604,6 +611,11 @@ public class AllayPlayer implements Player {
         }
     }
 
+    /**
+     * Adds type-specific metadata for a given entity to the provided {@code EntityDataMap}. Based on
+     * the type of the entity, this method updates the metadata to reflect the entity's specific state,
+     * attributes, and flags.
+     */
     protected void addTypeSpecificMetadata(Entity entity, EntityDataMap map) {
         switch (entity) {
             case EntityTnt tnt -> {
@@ -652,6 +664,9 @@ public class AllayPlayer implements Player {
                 if (attachedPlayer != null) {
                     map.put(EntityDataTypes.CUSTOM_DISPLAY, (byte) attachedPlayer.getRuntimeId());
                 }
+            }
+            case EntityEnderCrystal enderCrystal -> {
+                map.setFlag(EntityFlag.SHOW_BOTTOM, enderCrystal.isBaseVisible());
             }
             default -> {
             }